many-fixes

+++ init.sh
@@ -8,6 +8,8 @@ print() {
     echo -e "🔷 ${BLUE_BOLD}$*${RESET}"
 }
 
+cd $HOME
+
 print "passwd..."
 sudo passwd -d droid
 
@@ -17,12 +19,16 @@ sudo apt upgrade -y
 sudo apt install -y build-essential git
 
 print "git dotfiles..."
-cd $HOME
-git init
-git switch -c main
-git remote add origin "https://github.com/sandyberko/dotfiles"
-git fetch origin main
-git reset --hard origin/main
+if [ -d ".git" ]; then
+    echo "⚪ already exists - skipping"
+else
+    git switch -c main
+    git init
+    git remote add origin "https://github.com/sandyberko/dotfiles"
+    git fetch origin main
+    git reset --hard origin/main
+	git branch --set-upstream-to=origin/main main
+fi
 
 print "configure sshd..."
 sudo cp ~/.init_assets/sshd_config /etc/ssh/sshd_config
@@ -37,6 +43,14 @@ rustup default stable
 print "cargo binstall..."
 curl -L --proto '=https' --tlsv1.2 -sSf https://raw.githubusercontent.com/cargo-bins/cargo-binstall/main/install-from-binstall-release.sh | bash
 
+print "nushell..."
+cargo binstall -y nu
+command -v nu | sudo tee -a /etc/shells
+sudo chsh -s $(which nu) $USER
+
+print "configure vm..."
+sudo ~/.cargo/bin/nu ~/.init_assets/cfg_vm.nu
+
 print "jujutsu vcs..."
 cargo binstall -y jj-cli
 
@@ -46,17 +60,9 @@ curl -L --proto '=https' --tlsv1.2 -sSf \
 	| sudo tar xvzf - -C /usr/local/bin code
 
 print "vscode server service..."
-# Enable and start service
 sudo systemctl daemon-reload
 systemctl --user enable code-server
 systemctl --user start code-server
 
-print "nushell..."
-cargo binstall -y nu
-command -v nu | sudo tee -a /etc/shells
-sudo chsh -s $(which nu) $USER
-
-print "configure vm..."
-sudo ~/.cargo/bin/nu ~/.init_assets/cfg_vm.nu
-
 print "🎉 Done!"
+
